[feat] added route for bug reporting for specific apps added enum for application names added helper methods with type checking
